Bug: 解决uint64的遗留问题。解决mysql保存的时间不是东八区的问题。 Add: 集成Elasticsearch。可以在article_mq中订阅article表的数据，然后调用user_rpc集成usr表的数据。这样可以利用Es实现高效查询。

+++ application/applet/internal/logic/userinfologic.go
@@ -34,7 +34,7 @@ func (l *UserInfoLogic) UserInfo() (*types.UserInfoResponse, error) {
 		return nil, nil
 	}
 	u, err := l.svcCtx.UserRPC.FindById(l.ctx, &user.FindByIdRequest{
-		UserId: uint64(userId),
+		UserId: userId,
 	})
 	if err != nil {
 		logx.Errorf("FindById userId: %d error: %v", userId, err)
